+++ src/modules/edit/hooks/useNextAccessibleToilet.ts
