+++ src/modules/edit/tests/edit-wheelchair-description.e2e-spec.ts
@@ -1,12 +1,12 @@
 import { type Locator, expect, test } from "@playwright/test";
-import node4544823443Mock from "~/tests/e2e/edit/mocks/node-4544823443-osm-mock.json";
+import node4544823443Mock from "~/modules/edit/tests/mocks/node-4544823443-osm-mock.json";
 import {
   getButton,
   getDialog,
   getEditButton,
   getMenuItem,
   setupPage,
-} from "~/tests/e2e/edit/utils";
+} from "~/modules/edit/tests/utils";
 
 test.describe("Edit wheelchair description", () => {
   let dialog: Locator;
