+++ src/modules/edit/tests/mocks/way-126125230-osm-mock.json
