+++ src/needs-refactoring/components/CombinedFeaturePanel/components/AccessibilitySection/EditButton.tsx
@@ -3,7 +3,7 @@ import { t } from "@transifex/native";
 import { Pencil } from "lucide-react";
 
 import React, { useContext, useState } from "react";
-import { AutoEditor } from "~/pages/[placeType]/[id]/_components/AutoEditor";
+import { AutoEditor } from "~/modules/edit/components/AutoEditor";
 import { FeaturePanelContext } from "../../FeaturePanelContext";
 
 export function EditButton({
